+++ src/Config/packages.aget
@@ -2,13 +2,13 @@
   "nuget": {
     "references": {
       "DynamicLanguageRuntime": "1.2.2",
-      "DynamoVisualProgramming.Core": "3.0.0-beta6964",
-      "DynamoVisualProgramming.DynamoCoreNodes": "3.0.0-beta6964",
-      "DynamoVisualProgramming.DynamoServices": "3.0.0-beta6964",
-      "DynamoVisualProgramming.Tests": "3.0.0-beta6964",
-      "DynamoVisualProgramming.WpfUILibrary": "3.0.0-beta6964",
-      "DynamoVisualProgramming.ZeroTouchLibrary": "3.0.0-beta6964",
-      "DynamoVisualProgramming.DynamoSamples": "3.0.0-beta6964",
+      "DynamoVisualProgramming.Core": "3.0.0-beta7023",
+      "DynamoVisualProgramming.DynamoCoreNodes": "3.0.0-beta7023",
+      "DynamoVisualProgramming.DynamoServices": "3.0.0-beta7023",
+      "DynamoVisualProgramming.Tests": "3.0.0-beta7023",
+      "DynamoVisualProgramming.WpfUILibrary": "3.0.0-beta7023",
+      "DynamoVisualProgramming.ZeroTouchLibrary": "3.0.0-beta7023",
+      "DynamoVisualProgramming.DynamoSamples": "3.0.0-beta7023",
       "Greg": "3.0.0.3175",
       "GregRevitAuth": "2.0.8718.22924",
       "pythonnet_py38_win": "2.5.1",
